requirements.txt
torch>=2.1.0
torchvision>=0.16.0
opencv-python
Pillow
pandas
tqdm
numpy